+++ app/build.gradle.kts
@@ -41,8 +41,10 @@ kotlin {
             implementation(projects.core.ui)
             implementation(projects.core.domain)
             implementation(projects.core.data)
-            implementation(projects.core.datastore)
-            implementation(projects.core.remote)
+            implementation(projects.core.datasource.local)
+            implementation(projects.core.datasource.remote)
+            implementation(projects.external.network)
+            implementation(projects.external.datastore)
             implementation(projects.external.inappReview)
             api(projects.external.analytics)
             api(projects.external.crashlytics)
