+++ core/data/src/commonMain/kotlin/com/whatever/caramel/core/data/repository/AppRepositoryImpl.kt
@@ -2,6 +2,7 @@ package com.whatever.caramel.core.data.repository
 
 import com.whatever.caramel.core.data.mapper.toCheckForceUpdateResult
 import com.whatever.caramel.core.data.util.safeCall
+import com.whatever.caramel.core.datasource.local.LocalAppDataSource
 import com.whatever.caramel.core.datastore.datasource.LocalAppDataSource
 import com.whatever.caramel.core.domain.repository.AppRepository
 import com.whatever.caramel.core.domain.vo.app.CheckForceUpdateResult
