+++ core/datasource/local/src/commonMain/kotlin/com/whatever/caramel/core/datasource/local/LocalAppDataSource.kt
@@ -1,0 +1,16 @@
+package com.whatever.caramel.core.datasource.local
+
+import kotlinx.coroutines.flow.Flow
+
+interface LocalAppDataSource {
+    val activityParticipationCountFlow: Flow<Int>
+    val appLaunchCountFlow: Flow<Int>
+
+    suspend fun saveActivityParticipationCount(count: Int)
+
+    suspend fun saveAppLaunchCount(count: Int)
+
+    suspend fun saveInAppReviewRequestDate(dateTime: String)
+
+    suspend fun fetchInAppReviewRequestDate(): String
+}
