+++ core/datasource/local/src/commonMain/kotlin/com/whatever/caramel/core/datasource/local/LocalCoupleDataSource.kt
@@ -1,0 +1,9 @@
+package com.whatever.caramel.core.datasource.local
+
+interface LocalCoupleDataSource {
+    suspend fun fetchCoupleId(): Long
+
+    suspend fun saveCoupleId(coupleId: Long)
+
+    suspend fun deleteCoupleId()
+}
