+++ core/datasource/local/src/commonMain/kotlin/com/whatever/caramel/core/datasource/local/LocalTokenDataSource.kt
@@ -1,0 +1,14 @@
+package com.whatever.caramel.core.datasource.local
+
+interface LocalTokenDataSource {
+    suspend fun saveToken(
+        accessToken: String,
+        refreshToken: String,
+    )
+
+    suspend fun deleteToken()
+
+    suspend fun fetchAccessToken(): String
+
+    suspend fun fetchRefreshToken(): String
+}
