+++ core/datasource/local/src/commonMain/kotlin/com/whatever/caramel/core/datasource/local/LocalUserDataSource.kt
@@ -1,0 +1,9 @@
+package com.whatever.caramel.core.datasource.local
+
+interface LocalUserDataSource {
+    suspend fun fetchUserStatus(): String
+
+    suspend fun saveUserStatus(state: String)
+
+    suspend fun deleteUserStatus()
+}
