--- core/datasource/local/src/main/java/com/whatever/caramel/core/datasource/local/MyClass.kt
@@ -1,4 +1,0 @@
-package com.whatever.caramel.core.datasource.local
-
-class MyClass {
-}
