--- core/datastore/build.gradle.kts
@@ -1,19 +1,0 @@
-plugins {
-    id("caramel.kmp")
-    id("caramel.kmp.android")
-    id("caramel.kmp.ios")
-}
-
-android.namespace = "com.whatever.caramel.core.datastore"
-
-kotlin {
-    sourceSets {
-        androidMain.dependencies {
-            implementation(libs.koin.android)
-        }
-        commonMain.dependencies {
-            implementation(libs.bundles.datastore)
-            implementation(libs.koin.core)
-        }
-    }
-}
