--- core/datastore/src/androidMain/kotlin/com/whatever/caramel/core/datastore/DataStoreFactory.android.kt
@@ -1,10 +1,0 @@
-package com.whatever.caramel.core.datastore
-
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-
-fun createDataStore(context: Context): DataStore<Preferences> =
-    createDataStore(
-        producePath = { context.filesDir.resolve(DATA_STORE_FILE_NAME).absolutePath },
-    )
