--- core/datastore/src/commonMain/kotlin/com/whatever/caramel/core/datastore/DataStoreFactory.kt
@@ -1,13 +1,0 @@
-package com.whatever.caramel.core.datastore
-
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.PreferenceDataStoreFactory
-import androidx.datastore.preferences.core.Preferences
-import okio.Path.Companion.toPath
-
-fun createDataStore(producePath: () -> String): DataStore<Preferences> =
-    PreferenceDataStoreFactory.createWithPath(
-        produceFile = { producePath().toPath() },
-    )
-
-internal const val DATA_STORE_FILE_NAME = "caramel.preferences_pb"
