--- core/datastore/src/commonMain/kotlin/com/whatever/caramel/core/datastore/datasource/LocalCoupleDataSource.kt
@@ -1,9 +1,0 @@
-package com.whatever.caramel.core.datastore.datasource
-
-interface LocalCoupleDataSource {
-    suspend fun fetchCoupleId(): Long
-
-    suspend fun saveCoupleId(coupleId: Long)
-
-    suspend fun deleteCoupleId()
-}
