--- core/datastore/src/commonMain/kotlin/com/whatever/caramel/core/datastore/datasource/LocalTokenDataSource.kt
@@ -1,14 +1,0 @@
-package com.whatever.caramel.core.datastore.datasource
-
-interface LocalTokenDataSource {
-    suspend fun saveToken(
-        accessToken: String,
-        refreshToken: String,
-    )
-
-    suspend fun deleteToken()
-
-    suspend fun fetchAccessToken(): String
-
-    suspend fun fetchRefreshToken(): String
-}
