--- core/datastore/src/commonMain/kotlin/com/whatever/caramel/core/datastore/datasource/LocalUserDataSource.kt
@@ -1,9 +1,0 @@
-package com.whatever.caramel.core.datastore.datasource
-
-interface LocalUserDataSource {
-    suspend fun fetchUserStatus(): String
-
-    suspend fun saveUserStatus(state: String)
-
-    suspend fun deleteUserStatus()
-}
