--- core/datastore/src/iosMain/kotlin/com/whatever/caramel/core/datastore/DataStoreFactory.ios.kt
@@ -1,25 +1,0 @@
-package com.whatever.caramel.core.datastore
-
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import kotlinx.cinterop.ExperimentalForeignApi
-import platform.Foundation.NSDocumentDirectory
-import platform.Foundation.NSFileManager
-import platform.Foundation.NSURL
-import platform.Foundation.NSUserDomainMask
-
-@OptIn(ExperimentalForeignApi::class)
-fun createDataStore(): DataStore<Preferences> =
-    createDataStore(
-        producePath = {
-            val documentDirectory: NSURL? =
-                NSFileManager.defaultManager.URLForDirectory(
-                    directory = NSDocumentDirectory,
-                    inDomain = NSUserDomainMask,
-                    appropriateForURL = null,
-                    create = false,
-                    error = null,
-                )
-            requireNotNull(documentDirectory).path + "/$DATA_STORE_FILE_NAME"
-        },
-    )
