--- core/datastore/src/iosMain/kotlin/com/whatever/caramel/core/datastore/di/Module.ios.kt
@@ -1,11 +1,0 @@
-package com.whatever.caramel.core.datastore.di
-
-import com.whatever.caramel.core.datastore.createDataStore
-import org.koin.core.module.Module
-import org.koin.dsl.module
-
-actual val platformDataStoreModule: Module
-    get() =
-        module {
-            single { createDataStore() }
-        }
