+++ external/datastore/src/androidMain/kotlin/com/whatever/caramel/external/datastore/DataStoreFactory.android.kt
@@ -1,0 +1,10 @@
+package com.whatever.caramel.external.datastore
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+
+fun createDataStore(context: Context): DataStore<Preferences> =
+    createDataStore(
+        producePath = { context.filesDir.resolve(DATA_STORE_FILE_NAME).absolutePath },
+    )
