+++ external/datastore/src/commonMain/kotlin/com/whatever/caramel/external/datastore/datasource/LocalTokenDataSourceImpl.kt
@@ -1,0 +1,47 @@
+package com.whatever.caramel.external.datastore.datasource
+
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.core.edit
+import androidx.datastore.preferences.core.stringPreferencesKey
+import com.whatever.caramel.core.datasource.local.LocalTokenDataSource
+import kotlinx.coroutines.flow.first
+
+class LocalTokenDataSourceImpl(
+    private val dataStore: DataStore<Preferences>,
+) : LocalTokenDataSource {
+    override suspend fun saveToken(
+        accessToken: String,
+        refreshToken: String,
+    ) {
+        dataStore.edit { prefs ->
+            prefs[accessTokenKey] = accessToken
+            prefs[refreshTokenKey] = refreshToken
+        }
+    }
+
+    override suspend fun deleteToken() {
+        dataStore.edit { prefs ->
+            prefs[accessTokenKey] = ""
+            prefs[refreshTokenKey] = ""
+        }
+    }
+
+    override suspend fun fetchAccessToken(): String =
+        dataStore.data.first().let { prefs ->
+            prefs[accessTokenKey] ?: ""
+        }
+
+    override suspend fun fetchRefreshToken(): String =
+        dataStore.data.first().let { prefs ->
+            prefs[refreshTokenKey] ?: ""
+        }
+
+    companion object {
+        private const val PREFS_KEY_ACCESS_TOKEN = "accessToken"
+        private const val PREFS_KEY_REFRESH_TOKEN = "refreshToken"
+
+        private val accessTokenKey by lazy { stringPreferencesKey(PREFS_KEY_ACCESS_TOKEN) }
+        private val refreshTokenKey by lazy { stringPreferencesKey(PREFS_KEY_REFRESH_TOKEN) }
+    }
+}
