--- external/datastore/src/main/java/com/whatever/caramel/external/datastore/MyClass.kt
@@ -1,4 +1,0 @@
-package com.whatever.caramel.external.datastore
-
-class MyClass {
-}
