+++ settings.gradle.kts
@@ -38,7 +38,6 @@ include(":core:designsystem")
 include(":core:viewmodel")
 include(":core:ui")
 include(":core:remote")
-include(":core:datastore")
 include(":core:testing")
 include(":core:util")
 include(":feature")
